src/backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.16.1
python-dotenv==1.2.2
sqlparse==0.5.5
whitenoise==6.12.0
